+++ CHANGELOG.md
@@ -5,13 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.21.0] - 2024-09-07
 
 ### Changed
 
 - MAVLink extension now uses _decimal_ numbers for system and component IDs
   instead of hexadecimal. Using hexadecimal versions in earlier versions caused
   confusion and made the log messages harder to interpret.
+
+### Fixed
+
+- "Server location changed to unknown" message displayed no longer as it was
+  confusing during shutdown.
 
 ## [2.20.0] - 2024-09-03